backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.1
openai==1.3.5
python-dotenv==1.0.0
supabase==1.2.0
nltk==3.8.1
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
